pull out predicate and queryinterpreter
